projects/week7-ai-quality-monitor/requirements.txt
google-cloud-aiplatform
google-genai
python-dotenv